+++ src/array.rs
@@ -35,7 +35,6 @@ pub mod concurrency;
 pub mod data_type;
 mod dimension_name;
 mod element;
-mod endianness;
 mod fill_value;
 mod nan_representations;
 pub mod storage_transformer;
@@ -62,14 +61,15 @@ pub use self::{
     data_type::{DataType, DataTypeSize},
     dimension_name::DimensionName,
     element::{Element, ElementFixedLength, ElementOwned},
-    endianness::{Endianness, NATIVE_ENDIAN},
     fill_value::FillValue,
     nan_representations::{ZARR_NAN_BF16, ZARR_NAN_F16, ZARR_NAN_F32, ZARR_NAN_F64},
     storage_transformer::StorageTransformerChain,
 };
 pub use crate::metadata::v2::ArrayMetadataV2;
 pub use crate::metadata::v3::{fill_value::FillValueMetadata, ArrayMetadataV3};
-pub use crate::metadata::{chunk_shape_to_array_shape, ArrayMetadata, ArrayShape, ChunkShape};
+pub use crate::metadata::{
+    chunk_shape_to_array_shape, ArrayMetadata, ArrayShape, ChunkShape, Endianness, NATIVE_ENDIAN,
+};
 
 pub use chunk_cache::array_chunk_cache_ext_sync::ArrayChunkCacheExt;
 pub use chunk_cache::{