+++ src/metadata.rs
@@ -20,6 +20,7 @@ pub use array::{
 pub use group::{group_metadata_v2_to_v3, GroupMetadata};
 pub use v2::{ArrayMetadataV2, GroupMetadataV2, MetadataV2};
 pub use v3::{
+    codec::bytes::{Endianness, NATIVE_ENDIAN},
     AdditionalFields, ArrayMetadataV3, ConfigurationInvalidError, GroupMetadataV3, MetadataV3,
     UnsupportedAdditionalFieldError,
 };